move a bunch of code around
