+++ openmsipython/services/config.py
@@ -1,7 +1,7 @@
 #imports
 import pathlib, importlib, pkg_resources
 from inspect import isclass
-from ..utilities.logging import Logger
+from ..shared.logging import Logger
 
 class ServicesConstants :
     """
