+++ openmsipython/utilities/simple_provision_wrapper.py
@@ -1,9 +1,9 @@
 #imports
 import pathlib, shutil, logging, urllib.request, kafkacrypto
 from argparse import ArgumentParser
+from ..shared.logging import Logger
+from ..shared.config_file_parser import ConfigFileParser
 from .misc import cd
-from .logging import Logger
-from .config_file_parser import ConfigFileParser
 
 #constants
 LOGGER = Logger('ProvisionNode',logging.INFO)
