+++ test/unittests/test_serialization.py
@@ -1,12 +1,12 @@
 #imports
-from config import TEST_CONST
+import unittest, pathlib, logging
+from confluent_kafka.error import SerializationError
+from openmsipython.shared.logging import Logger
 from openmsipython.my_kafka.serialization import DataFileChunkSerializer, DataFileChunkDeserializer
 from openmsipython.data_file_io.upload_data_file import UploadDataFile
 from openmsipython.data_file_io.data_file_chunk import DataFileChunk
 from openmsipython.data_file_io.config import RUN_OPT_CONST
-from openmsipython.utilities.logging import Logger
-from confluent_kafka.error import SerializationError
-import unittest, pathlib, logging
+from config import TEST_CONST
 
 #constants
 LOGGER = Logger(pathlib.Path(__file__).name.split('.')[0],logging.ERROR)
