+++ test/unittests/test_upload_data_file.py
@@ -1,8 +1,8 @@
 #imports
 import unittest, pathlib, logging
 from queue import Queue
+from openmsipython.shared.logging import Logger
 from openmsipython.data_file_io.config import RUN_OPT_CONST
-from openmsipython.utilities.logging import Logger
 from openmsipython.data_file_io.upload_data_file import UploadDataFile
 from config import TEST_CONST
